requirements.txt
streamlit
langchain
langchain-ollama
langchain-chroma
pywebview